Updated openstack/openstack Project: openstack-infra/python-storyboardclient  c0a604b432a8769c208d873eff40e218aeb1a980
Update Docs

The installation instructions were incorrect; they advised the user to
install a package via pip that was not registered in pypi. This patch
changes the installation instructions to ones that work. To avoid
confusion, this patch also notes that the client does not run via a
command, but can just be imported into other scripts.

Change-Id: If34a26dc78641daf0be90dab4a7f788dc72db4bb