Updated openstack/openstack Project: openstack-infra/tripleo-ci  1b316d44a51be68bbe9cc49e087cd21a88a15dc7
Don't execute the output from wait_for

This if statment is intended to simply run wait_for, we don't want
to run the output from the command. If then event an error occurs
wait_for outputs a message about a timeout. We want to see this message
not execute it on the command line.
/tmp/tripleo-common/scripts/tripleo.sh: line 509: Timing: command not found

Change-Id: I1ee1572cb8720ae442dfbe4a81e5c02118cdf803