Updated openstack/openstack Project: openstack/cinder  3fc0c10e71d81d4cd0b805fe78bcd7a751e1013c
VNX: Add 'None' check in parsing provider location

Manage existing volume that requires a migration in VNX raises error in
VNX driver.

The root cause is that VNX uses a function to retrieve volume type from
provider location.  When provider location is not initialized (such as a
new LUN), it will send `None` to the VNX driver and driver didn't handle
it well.

Add condition check for this function to return `None` if `None` is
received.

Closes-Bug: 1516508
Change-Id: I71bff4f896aa204d10502b01ec8259d74ee4a523