Updated openstack/openstack Project: openstack-infra/shade  75ee8f967df00a8595c8169a1958295839aae387
Mutex protect internal/external network detection

We have stamp variables to prevent double processing the events as a
mechanism of caching, but they're not mutex protected, so we can also
have a race-condition thundering herd potential in multi-threaded code
like nodepool. Add mutex protection around it - but just use one mutex
for the two, because we can further refactor this to make a single pass
through the network list and sort the networks from that.

Change-Id: I74d8b279105e8d9e1a2c138da90e31485e123744