Updated openstack/openstack Project: openstack/keystone  2dddbfeb5cb7f16ebc6214992427aeae321f8850
Correct test_implied_roles_fk_on_delete_cascade

This test was using the role_api to insert rows into and query the
database. The rows inserted into the database for migration tests
must be inserted directly because the driver is not necessarily
going to work with the old version of the database that's in use in
the test and the contents of the database could change and not match
what the old version of the code actually inserted.

Change-Id: I3e05e132139623c6323338c3ba5a0742542e7562