Updated openstack/openstack Project: openstack/openstack-manuals  391728d3218dd7600ae6c44a19ad76525e5931ac
[user guide] Update text references on links

In the End User Guide, updated the link text for
references to the Cloud Admin Guide to Admin Guide.
As part of reorg, the Cloud Admin Guide title will be
changed to Admin Guide. This review is only part of
the work required to rename the guide for Mitaka.

Change-Id: Ic491225bf89b5793f0effaf4d78dd7b44a16bcca
Implements: blueprint user-guides-reorganised