Updated openstack/openstack Project: openstack/neutron  f3be18a758180e317050062125544d81b32681f6
Remove obsolete keepalived PID files before start

keepalived refuses to start and claims "daemon already started"
when there is already a process with the same PID as found in
either the VRRP or the main process PID file. This happens even
in case when the new process is not keepalived. The situation
can happen when the neutron node is reset and the obsolete PID
files are not cleaned before neutron is started.

This commit adds PID file cleanup before keepalived start.

Closes-Bug: 1561046
Change-Id: Ib6b6f2fe76fe82253f195c9eab6b243d9eb76fa2