Updated openstack/openstack Project: openstack/governance  63839adc67682d3b66179396b6969679d9f13710
Update Kuryr mission statement

Kuryr currently bridge between containers networking models
(Like Kubernetes/Docker and others) to Neutron and OpenStack
Networking advance services.

When one think about this model, storage and backup areas
in Openstack are pretty mature and flexible to enhance containers
in terms of storage and volume sharing/re-cycling and more.
(Not to mention backups which there are already few projects in
OpenStack in that area)

Our team agreed that Kuryr seems the natural place to expand
our concept to storage just as we do right now for networking.
Its important to note that this doesnt overlap with any OpenStack
project that we currently know of.

Change-Id: Ie94dcd581ab004b38d1112d784df8f527dea5f10