Updated openstack/openstack Project: openstack-infra/project-config  8e6535ceca291c776ed0b6f02aacb0e1dcd09e8e
Remove oslo.messaging from bandit integration

This patch temporarily removes oslo.messaging from the bandit
integration until that project has bandit passing as part of its
gate. The reference patch for this is:
    https://review.openstack.org/#/c/286506/

In this way, we can proceed with a new release of bandit since
all other integrations are passing.

Change-Id: I13076dd36f8b0b584d5289c217105d0bd75207f2