Updated openstack/openstack Project: openstack-infra/project-config  d7dce6bc5407970e71a9ba8d459b7eb0c179ee7e
devstack-plugin-nfs: Make tempest non-voting

This is to resolve a chicken/egg problem with
this plugin.  The NFS code in Cinder is not yet in shape
to pass all tempest tests, and I'd like to use this
plugin to vet that code while it's in-progress with a
non-voting Cinder job.

So, allow commits to the plugin for now until everything
is up and running, and then this job can be made voting
again.

Change-Id: Ia01d37951d4ba49da6a55e90d8379f3a29e6975c