Updated openstack/openstack Project: openstack-infra/project-config  a4107ae166ce778ece742c4086bdc9e2d193327b
fuel-stats: use db setup for python jobs

Use the python db jobs. These run the same tests as the normal jobs,
just setup databases. This is in preparation for a future cleanup of
the default python jobs so that they those not run databases by default.
These repo need databases setup for tests.

Change-Id: Ib06c9c94cb767e790d6420093244d722a1b1eda3