Updated openstack/openstack Project: openstack/horizon  e27513d11cc2c7358c1ac9740de572896b4c044a
Duplicate help message in metadata tab

When launching an instance or updating the metadata of an image,
there is a duplicate help text on the top and the bottom of the
metadata tab.
This patch will change the text on the bottom by a meaningful
message to avoid the repetition.

Change-Id: Idc41db67edaa5bb2921daf958d1e7bf8a07aa648
Closes-Bug: #1558002
Closes-Bug: #1563137