Updated openstack/openstack Project: openstack/tripleo-heat-templates  cccb984eb76febfedfe7118e27d3d82f31a224e2
compute: restart OVS agent on neutron.conf changes

Some options in neutron.conf are used bu OVS agent, like logging &
messaging.
During the upgrade process, you need to restart the agent if these
options change.

We could patch puppet-neutron to add a notify, but the community won't
like it because Neutron OVS agent is not able to restart gracefully
until [1] got merged. Until that, we can fix it in TripleO, where we
suppose Puppet runs happenning during bootstraps and upgrades.

Later, we'll drop this code from here and move it in puppet-neutron.

[1] https://review.openstack.org/#/c/297211

Change-Id: I02b17b66e93331ddfb1a7abd8adff672bc7a32d6
Closes-Bug: #1563437