Updated openstack/openstack Project: openstack-infra/project-config  3590ff967d4e944893b4486307eb684f92c568f9
Add job to test Nova with specific libvirt and kvm packages

Nova has sometimes the need to exploit features of libvirt which
are not yet in the default libvirt version of our testing
infrastructure.
This change adds a test job which is able to set up a devstack node
which runs with specific libvirt and kvm packages.

Change-Id: Ifd231857cffd134e911225088d6dc65bf61fdedd