Updated openstack/openstack Project: openstack/python-designateclient  d89f0a3889a8d258b2d3d73c6ed202f6fc9cafc3
Add a service catalog override to the functional tests

This allows us to run the functional tests against environments than
what's in the service catalog.

note: ideally, this would allow us to run against a Designate in
noauth mode. And it does, however there's not a way to pass in the
X-Auth-Project-ID header, so the tests will always use the
noauth-project (tests involving multiple users like, zone transfer
tests, won't work).

Change-Id: If9306b27d78d7c774911d3c1abb46bfbf72539f1