Updated openstack/openstack Project: openstack/keystone-specs  5d6ecd7566627db93d963a29f2d4248b1542e0c7
Shadow users (continuation for newton)

We completed the groundwork detailed in the original shadow users spec
during Mitaka; the work remaining for Newton is to begin to take
advantage of the newly unified data model.

This spec carries over the remaining work from Mitaka and only details
the differences from the original spec.

Change-Id: I475c365f528a536ac289cfb86a14519e5b3bd1bc