Updated openstack/openstack Project: openstack-infra/project-config  3db39864fd2c251ede88d187b83c3ad394c65555
propose-requirements-updates should run on all branches

This job picks up whatever is in global-requirements in a
specific branch and files reviews in projects that need
an update. So this should run on all branches, not just
master. This was broken in the following change:
Iefec39861421a77643e0c06acc36689ac863e2ef

Change-Id: I731eee2b23cf3d0ba6ff37c4f501d285f020fdb1