Updated openstack/openstack Project: openstack/heat  7933a8bc3d8ad4a97363e89707a74e3c67988640
Don't validate template resource parameter defaults in parent

When creating a properties schema for a TemplateResource, we should not
include the default parameter values specified in the template because
properties lack the type conversions that parameters do (specifically, a
comma_delimited_list parameter converts a string value to a list using
str.split()). We don't even pass defaults to the nested template anyway,
because it is the parameter parsing in the child stack that handles
substituting in any defaults.

Change-Id: Ic7251cf0fcb77cc2acae43dc3c1d9f4671e4a52e
Closes-Bug: #1564106