Updated openstack/openstack Project: openstack-infra/project-config  fc9922f93a7f85f46014b8067bf845f678a4c413
Revert "move fedora23 devstack to experimental"

This reverts commit 2b2489966622939f17fe2250678abe9efac761eb.

The issue of not running on the right performance nodes was corrected
with I64cf074424b3c8eebf67510ba91bf24f2680dd0d.  I have done a number
of runs since then, all of which have landed on a RAX node

 [1] 50m 59s
 [2] 59m 02s
 [3] 1h 01m 44s

This low number is about what I would expect; it is about 5 minutes
more than other platforms running the same tests and is mostly taken
up by building pip wheels.

Note; we are considering a wheel builder for Fedora.  In terms of the
building, mirror and job setup, that's all quite straight forward.
The only problem might be with the infra puppet that sets up slave
nodes.  Fixing this for new releases is generally overall useful work;
it's not like the new stuff Fedora ships is going away (e.g. see
puppet4 efforts), but it can often be some effort to get this going.
This means we can not really rely on a pip mirror being up the same
day as we want to be testing devstack.

[1] http://logs.openstack.org/62/295062/1/experimental/gate-tempest-dsvm-platform-fedora23-nv/06bb66f/
[2] http://logs.openstack.org/62/295062/1/experimental/gate-tempest-dsvm-platform-fedora23-nv/1b21f93/
[3] http://logs.openstack.org/62/295062/1/experimental/gate-tempest-dsvm-platform-fedora23-nv/2941b3e/

Change-Id: I50048b347693b6743849220458dab4756a4a1c71