Updated openstack/openstack Project: openstack/python-openstackclient  e666a8ab74bff25d3a5d9897618bc08c4af91359
Aggregate object should be "grouping of compute hosts"

The describe of Compute "aggregate" in command-objects and commands
documents use "a grouping of servers", but exactly that should be a
grouping of compute hosts. "server" object in OSC is a virtual machine
instance, and server group should be "grouping of servers".

Change-Id: Ib034fed15f11fc3e756985b3131a9922129ed6bf
Closes-Bug: #1563172
Related-Bug: #1542171