Updated openstack/openstack Project: openstack-infra/project-config  90bc24b54fa01699ba3247044ceccbaf40833536
Add removal of grub boot timeouts back in

When we removed the restrict memory grub stuff because hpcloud was no
longer in use we also removed the configuration updates that set the
grub timeout to 0. This can save many seconds per instance boot purely
waiting for grub to timeout and continue booting the host.

Change-Id: I2d79ec4ebd55f927d868e06c8f00cf50f34f3b92