Updated openstack/openstack Project: openstack-infra/devstack-gate  eaed7eca6c56594d6312233a57ab9c0adf985d88
Save PROJECTS form the job defintion for use in reproduce.sh

If the Jenkins Job definition adds items to the PROJECTS variable that
are not included as we build up PROJECTS reproduce.sh will ultimately
fail.

The fix saves the PROJECTS as it is passed into devstack-gate-wrap and
includes that in reproduce.sh.

Change-Id: Iea45f53b8c26985cb0da9720a631fcf8725522cf
Closes-Bug: bug/1544827