Updated openstack/openstack Project: openstack/keystone  ca5074bed546ec5f6a8ffdab4d8c3655bed8130a
Improve error message for schema validation

Without this patch, users who enter long project names that fail the
jsonschema validation are faced with the message:

  Invalid input for field 'name'. The value is '<very long name>'.

which simply echoes back the input value and does not provide
information about what was wrong with the input. This is less than
helpful. This patch modifies the validator error message to, in
addition to supplying the name of the failed resource, also provide the
error message directly from the jsonschema exception. In apparently all
cases this message includes the value the user entered, so we can omit
echoing the value back. An example of the new message is:

 Invalid input for field 'name': u'<very long name>' is too long

An alternative to this patch would be to follow the pattern given by
the "create_user" action, where jsonschema does not validate the
maxLength and it is instead validated at the core level by a
keystone.common.clean function. This would solve bug 1557609 and give
us a more descriptive error message, but moves validation further down
the chain and contradicts bp api-validation.

Change-Id: I3e5a41be91ebe87f33eab9e545f330dc141bb88e
Closes-bug: #1557609