Updated openstack/openstack Project: openstack/cinder  238e3eb1f4906f6c122417b57a0341b6853b57d3
Storwize: fix deadlock and major lock granularity issue

Change the granularity of the lock from all Storwize backends to locking
based on "Stowize system id and host" the volume is being mapped to.
This granularity will prevent mappings to the same "host"
at the same time on the same Storwize system so that the
first initialize_connection will create the host entry and the last
terminate_connection will remove the host entry. This granularity of
lock properly protects the mapping and un-mapping of the volume
to the host entry on the targeted SVC. Move the lock decorator from
initalize_connection and terminate_connection to a locking methods
_do_initalize_connection_locked and _do_terminate_connection_locked
which in turn call _do_initalize_connection and _do_terminate_connection.

From _do_initalize_connection on a failure call _do_terminate_connection
as _do_initalize_connection already holds the lock calling
terminate_connection results in a deadlock.

Fix a bug in get_host_from_connector that is exacerbated by the
new locking granularity.  If a host happens to have been removed
via a terminate_connection or storage administrator action
between the time the method starts and the time we query
to get the details on the host entry in the list,
we now properly ignore the failure to find that object
and continue on to the next entry.

Add UT to cover new code paths.

Change-Id: I66b8f1f8aeefbfebeb9acab247c1a0254ff5f618
Closes-Bug: #1557822