Updated openstack/openstack Project: openstack/neutron  2a7b642d1d96a15a57f7039f19c5f2dd546b7f81
Refactor and fix dummy process fixture

Extracting the test fixture that creates a new process and leaves it
running for a given amount of time into helpers where other fixtures for
functional tests live. This both keeps the fixtures at one place and
increases visibility of the fixture so that it can be reused in other
tests. At the same time, the fixture is fixed as the original code
omitted starting the process.

Change-Id: I97aeb8d1d5773ef3d59e8f908aea34ccceb38378
Related-Bug: 1561046