Updated openstack/openstack Project: openstack-dev/devstack  5c775831b3db8b9575e433d924807aa2702b5393
Move auth setup earlier in tempest config

This commit moves the auth setup for tempest config to occur before
we run tempest verify-config. The API requests that command runs
require auth and in the case we run tempest without admin creds set
the config file will not have any credentials to run the query with.
By moving the auth setup to occur before this it will ensure tempest
is always configured with credentials before we run the command.

Change-Id: I6d11b24e4492f1fde3aa3a7a239c40d63111bfa1