Updated openstack/openstack Project: openstack/cinder  f0a81d5abc26eb34bdbbb9cedb2db5a1373c10f1
Re-enable image to volume with failure test

The unit test test_copy_volume_to_image_with_image_volume_failure
was being skipped because the volume availability_zone attribute
(and a few others) were non-nullable and were not getting set. As of
7a158ac4488ac86c742e15c714dda7dd76772366 these attributes are nullable
and the tests now succeed.

The remaining tests were fixed to use valid id's for the volumes so
that warnings aren't emitted when running the test case.

Closes-Bug: #1525773
Change-Id: I6ee4f878d70d12e64f0c01807c58900bc48c0173