Updated openstack/openstack Project: openstack-infra/tripleo-ci  76adb7d9af748651908b276a240a15a0f5684e94
Clarify location of large output log files

Currently we log some messages that tell the user to look in
/var/log for output from the undercloud install and image builds.
However, we don't tell them where that /var/log actually is.  In
this case it's in undercloud.tar.xz in the CI job directory, so
this change adds that information to the message.

Change-Id: Icb1fd2da06fa8ced9f6a591df88f53c018679398