Updated openstack/openstack Project: openstack/neutron  564a79abb117a5565588a40d46e7e95eddf5b54e
Rename first argument to extend_dict hook correctly

The first argument should not be called 'self', since the method
is a module function. It should more aptly called 'core_plugin',
because this is what the object passed to the method is going to be.

This is an innocent oversight, but stylistically wrong nonetheless.

Change-Id: Iedc2f227e6875ead0085de0bd3b00ad0477abc83