Updated openstack/openstack Project: openstack/governance  d0bddd89fbcfc166d04a5bb86acc6fa4ddf2e5a8
Update some tags for the Trove project

This is the first of multiple changes to fix tags on the trove
project.

In reviewing the tags for the Trove project, I find some tags and
assertions to be missing and I have added them. I have run the
validate-tags tool and it reported no errors relative to Trove. (to
save time, I ran the tool with a modified projects.yaml file that had
only Trove). I have retained the release:has-stable-branches on
trove-dashboard as that is the intent.

Change-Id: Ide690dbae0539326d287a57856eeef60e8ce29e1