Updated openstack/openstack Project: openstack/governance  f4bd80227336f18d5e812f3cbb55f5d145c285b0
Create puppet-openstack-guide

Currently Puppet OpenStack documention is hosted on Wiki [1] and we would
like to use RST and Gerrit to manage documentation in future, like other
projects, so we can leave Wiki.

On each existing Wiki page of Puppet OpenStack, we'll give the new URL
so we don't break indexes.

[1] https://wiki.openstack.org/wiki/Puppet
Change-Id: I8a42fd763913775834c9e621ca7582199705773d