Updated openstack/openstack Project: openstack/heat  d07845c2a15562b153f74d6c62c6c7b396bc67fa
Avoid reference loops in DependencyTaskGroup

Holding a reference to a traceback in a local variable can cause a
reference loop (since the stack frame containing the local variable is
inevitably referenced by the traceback), with the result that basically
everything may need to get garbage collected. Prevent this by explicitly
deleting the local variable holding the deletingtraceback once it is no
longer required.

Change-Id: Icde3dc0c6d5b2c799df778d3e1780514bc04bf5b