Updated openstack/openstack Project: openstack/nova  f95332ab491c43394dd87b9d9c3c7d440be2aa29
neutron: Update the port with a MAC address for PFs

In case of an SR-IOV PF as a neutron port, we should really update
Neutron with the proper MAC, as it is known only once the instance has
claimed a device on the compute host.

This patch adds a local Neutron API method that will grab the MAC
address based on the claimed device, and pass it on with the binding
port updates that happen during network allocation.

Change-Id: I7eafcaa9147bb9cbe7c7c8cda34ec259be0b0eb5
Implements: blueprint sriov-pf-passthrough-neutron-port