Updated openstack/openstack Project: openstack/nova  c368a3e802a9b9f2ddabcc2660fecf3615d9920c
The 'record' option of the WebSocketProxy should be string

If proxy service (VNC, spice, serial) is started with the 'record'
option, an exception is thrown. This is because WebSocketProxy
expects a file path for the record argument instead of boolean
value. This patch fixes this by changing 'record' to be string
option. An empty string means no recording.

Closes-Bug: #1559026

Change-Id: Id6cf7c65191bdaa579f7cdadea43230897436de6