Updated openstack/openstack Project: openstack-infra/shade  79b458414e8ed045fde485380b6b75268973cfe8
Remove conditional blocking on server list

When the server list cache proceduce was imported from nodepool,
it grew this internal conditional which caused all server list
calls to block on aquiring the server list cache update lock iff
the server list cache is empty.  This means that on startup, we
could have a thundering herd of threads which all insisted on
waiting for the lock because there was no cached server list.  This
does not seem to be necessary, and the original nodepool logic of
always performing non-blocking acquisitions of the lock and
returning the old (possibly empty) list of servers for any thread
that did not get the lock (but will try again a few seconds later
and should receive updated data at that time) should be sufficient.

Change-Id: I5640f60da2b7789a98bea033e16695389c6062e0