Updated openstack/openstack Project: openstack-infra/shade  75bb222fddbfeb30097b5b97b5225fe6e024d1b3
Make get_stack fetch a single full stack

There is a convention in shade for get_<entity> calls to do a
list_<entities> with no server filtering then do a client-side filter
to return a single result. This approach is not appropriate for Heat
since the stacks.list call returns significantly less detail than the
stacks.get call.

A user calling get_stack will almost certainly be wanting the extra
detail provided by the proper stacks.get (for example, the stack
outputs).

This change switches to using stacks.get for the get_stack
implementation. It replaces the search_stacks call with a local
function that returns a 'list' of one full stack object.

Change-Id: I5d326d489f806709252a22360f3dbd8011fdb9c7