Updated openstack/openstack Project: openstack/api-site  16027a615b1399e4654f5ae311efd35f18ce09a3
Delete "pool_id" in JSON request format in Create a health monitor

The request parameter shown in JSON request format is not listed
in request parameters table and is not a legal attribute when
creating a load balancer health monitor, thus should be deleted.

The corresponding JSON response format is also fixed.

Change-Id: Ide050fbbce898cdd58785840f65c75bbb42db5a2
Closes-Bug: #1564227