Updated openstack/openstack Project: openstack/api-site  ea34f1d4b19372315ce13f94db4d42f8f22e7542
Remove passwordResponseParameter from Identity V3 API

Retrieved user details do not include the user password, therefore removing this
entry. The password is filtered out [1] and not returned when accessing the API
using admin credentials.

There was also a passwordResponseParameter entity defined but not used in the
database section, removed that entry as well.

[1] http://git.openstack.org/cgit/openstack/keystone/tree/keystone/identity/core.py#n59


Closes-Bug: 1564768
Change-Id: I82f2a55bf67a0a5a0fb1d1499ca48a507252dd2a