Updated openstack/openstack Project: openstack/nova  d233772b0d6880fe6dfe3e28125e47ff1a467ee6
Fix: unable to delete instance when cinder is down

Ignore any cinder related exceptions that arise when you try to delete
an instance, so that the delete instance workflow isn't abruptly
interrupted (skipping necessary clean-up, etc).

In this case, we are intentionally catching all exceptions in addition
to a specific set of exceptions, so that the instance and related
resources aren't left in an unmanageable state when an unexpected
exception occurs. For example, currently you can't delete an instance
that's attached to a volume when cinder is down.

The existing error messages were also updated to include the block
device mapping volume id that triggered the exception.

Change-Id: Iba7b4cc4b59f88b0817c4618e7a4429161d6c2a9
Closes-Bug: #1563547