Updated openstack/openstack Project: openstack/nova  18169d572a784e5ca795b09415631a73743f12c9
Enhanced error handling for rest_parameters parser

Instead of throwing a KeyError if the lookup file isn't what we
expected, and dying, throw a warning and skip the key in content. This
makes it a bit easier to understand what's wrong and how to address it
for people not familiar with the code.

Change-Id: I0d1cb60b5a1010d00e5290f697bbde601fd5f221