Updated openstack/openstack Project: openstack/python-heatclient  cc4a611b7bf4b94121976b602095d36ee3d75fec
OSC plugin should be region/interface aware

This fixes the OSC plugin so that if region is specified or if the
interface is specified (public, internal, admin) then those will be
passed into the heat_client when instantiated.  Without this the
client will pick the endpoint for the first region returned instead of
the one specified.

Closes-Bug: #1556263
Change-Id: I9c15719e7711c93e7fd076d38d66011bb20f988d