Updated openstack/openstack Project: openstack-infra/system-config  19051f17fdf70e5f468740ad680f372122b34051
Declare redis 2.8.4 on openstackid.org

We're upgrading openstackid.org to Ubuntu 14.04 LTS, and so need to
adjust the redis version passed to the openstackid module for it.
This matches what's already in use for openstackid-dev.openstack.org
and Puppet is currently disabled on the real openstackid.org by UUID
while we build its replacement.

Change-Id: I8b628446736e296aa6028c8e2ec848f2db8c4373