Updated openstack/openstack Project: openstack/tripleo-heat-templates  9926cfb5324e2ac1e226bf7a3d5650a823d166ab
Nova needs the proper volumes to use Cinder

Kolla has been using ceph.  For a while, cinder had
iscsi build into it, but it was removed.  In order to
get this to work with containers again, nova-compute and
libvirt containers need /dev and /lib/udev mounted into their
containers.

We also need to copy nova's rootwrap.conf into the nova.
It was missing this config file.

Change-Id: Ie77f56b4576d5393ad3756b0f5ecc3eeff844d1f