Updated openstack/openstack Project: openstack/tempest  eda607cb1c89762a54290c4525d81c1b192c421f
Remove AuthorizationTestJSON tests

These are a set of legacy negative tests that make an assumption that
Nova service URLs will have {project_id} in them, and that changing
the value of {project_id} is a meaningful way to attempt to access
resources by another tenant. Neither of these things are true.

As of midway through Mitaka Nova landed a microversion that makes
{project_id} an optional construct. When we actually deploy systems
with this, of the 25 tests listed, 20 fail, 3 are skipped (due to
custom skip logic I previously add), and 2 pass. Those 2 passes are
actually as suspect as anything else, as they should now fail, and
demonstrates other bugs in Tempest.

The removal of these tests followed the full process required for
removal of tests -
http://lists.openstack.org/pipermail/openstack-dev/2016-March/088716.html

Change-Id: I81d99d65b8affaea44ddd6643d6acc40068eead7