Updated openstack/openstack Project: openstack/requirements  622f3985500e0b7f74adac4771033e5d3e1e5ed3
New library split from Magnum - python-k8sclient

In I1472e631ec7f06e74e73c27edd68a1e2dc7b23e1, we created
a new repo to hold the python code that talks to
kubernetes. In this review, we are add the first test
release 0.1.0 to enable Magnum to remove the code
in Magnum with this new library.

Change-Id: Ia400b5b6919753d6241233a6666a2c30c7618bc9